+++ src/main/java/Deadline.java
@@ -2,7 +2,7 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
 
     protected String deadline;
     protected LocalDate deadlineTime;
