+++ src/main/java/Event.java
@@ -2,7 +2,7 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Event extends Task{
+public class Event extends Task {
 
     protected String time;
     protected LocalDate eventTime;
@@ -17,7 +17,7 @@ public class Event extends Task{
     Event(String description, boolean isDone) {
         super(description, isDone);
         this.time = "";
-        this.eventTime=null;
+        this.eventTime = null;
     }
 
     Event(String description, String time) {
