+++ src/main/java/ToDo.java
@@ -1,4 +1,4 @@
-public class ToDo extends Task{
+public class ToDo extends Task {
     ToDo(String description) {
         super(description);
     }
