+++ src/main/java/Ui.java
@@ -1,0 +1,2 @@
+public class Ui {
+}
